requirements.txt
aniso8601==9.0.1
category-encoders==2.5.1.post0
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
Flask==2.2.2
Flask-API==3.0.post1
Flask-RESTful==0.3.9
idna==3.4
imbalanced-learn==0.10.0
imblearn==0.0
importlib-metadata==5.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
MarkupSafe==2.1.1
numpy==1.23.5
packaging==22.0
pandas==1.5.2
patsy==0.5.3
psycopg2-binary==2.9.5
python-dateutil==2.8.2
pytz==2022.6
requests==2.28.1
scikit-learn==1.2.0
scipy==1.9.3
six==1.16.0
statsmodels==0.13.5
threadpoolctl==3.1.0
urllib3==1.26.13
Werkzeug==2.2.2
zipp==3.11.0
